D20 Adventure- 06/23/22
Welcome to the dungeons! There is wealth and fame abound here however it won't come easy (and it never does, does it?). A monster bars your path and you must slay it to continue forth! Or will you fall to this unsightly beast! Roll your trusty d20 to find out